requirements.txt
psycopg2
configparser
Pillow
